# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api SDE for mobile sde",
        "description": "Implementation of Swagger with in Laravel",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://haki44.eu/api",
            "description": "API SDE"
        }
    ],
    "paths": {
...
        "/skipper/{id}": {
            "get": {
                "tags": [
                    "Skippers"
                ],
                "summary": "Show one skipper",
                "description": "Show one skipper",
                "operationId": "ShowSkipper",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of skipper",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Skippers"
                ],
                "summary": "Update one skipper",
                "description": "Update one skipper",
                "operationId": "UpdateSkipper",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of skipper",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Skippers"
                ],
                "summary": "Delete one skipper",
                "description": "Delete one skipper",
                "operationId": "DeleteSkipper",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of skipper",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    }
}